requirements.txt
transformers
sentence-transformers
Pillow
faiss-cpu
torch
fastapi
uvicorn
opencv-python
scikit-learn
pandas
streamlit
scipy
matplotlib
seaborn
IPython
requests
numpy
beautifulsoup4
python-multipart